// src/types/spotify.d.ts
/* eslint-disable no-use-before-define */
// Generated by https://quicktype.io

export interface NowPlaying {
  timestamp: number
  context: Context
  progress_ms: number
  item: Item
  currently_playing_type: string
  is_playing: boolean
}

export interface TopTracks {
  items: Item[]
}

export interface Context {
  external_urls: ExternalUrls
  type: string
  uri: string
}

export interface ExternalUrls {
  spotify: string
}

export interface Item {
  album: Album
  artists: Artist[]
  duration_ms: number
  external_urls: ExternalUrls
  name: string
  preview_url: string
  type: string
  uri: string
}

export interface Album {
  album_type: string
  artists: Artist[]
  external_urls: ExternalUrls
  images: Image[]
  name: string
  type: string
  uri: string
}

export interface Artist {
  external_urls: ExternalUrls
  name: string
  type: string
  uri: string
}

export interface Image {
  height: number
  url: string
  width: number
}

/**
 * Response Data
 */

export interface NowPlayingResponse {
  isPlaying: boolean
  from: string
  fromUrl: string
  name: string
  url: string
  artists: Array<{
    name: string
    external_urls: string
  }>
  smallImage?: string
  bigImage?: string
}

export interface TopTracksResponse {
  items: Array<{
    name: string
    url: string
    artists: Array<{
      name: string
      external_urls: string
    }>
  }>
}